refactor: useReportMessage 훅에서 checkDuplicateReport 및 reportMessage로 서비스 함수 분리
